Rename ChoiceMenu into ButtonMenu, move to package
