+++ src/main/kotlin/io/github/freya022/botcommands/api/pagination/menu/buttonized/ButtonMenu.kt
@@ -1,8 +1,9 @@
-package io.github.freya022.botcommands.api.pagination.menu
+package io.github.freya022.botcommands.api.pagination.menu.buttonized
 
 import io.github.freya022.botcommands.api.components.event.ButtonEvent
 import io.github.freya022.botcommands.api.core.BContext
 import io.github.freya022.botcommands.api.pagination.Paginators
+import io.github.freya022.botcommands.api.pagination.menu.AbstractMenu
 import net.dv8tion.jda.api.interactions.components.buttons.ButtonStyle
 import net.dv8tion.jda.api.utils.messages.MessageCreateBuilder
 
@@ -11,12 +12,12 @@ import net.dv8tion.jda.api.utils.messages.MessageCreateBuilder
  *
  * @param E Type of the entries
  *
- * @see Paginators.choiceMenu
+ * @see Paginators.buttonMenu
  */
-class ChoiceMenu<E> internal constructor(
+class ButtonMenu<E> internal constructor(
     context: BContext,
-    builder: ChoiceMenuBuilder<E>
-) : AbstractMenu<E, ChoiceMenu<E>>(
+    builder: ButtonMenuBuilder<E>
+) : AbstractMenu<E, ButtonMenu<E>>(
     context,
     builder,
     makePages(builder.entries, builder.transformer, builder.rowPrefixSupplier, builder.maxEntriesPerPage)
