+++ src/main/kotlin/io/github/freya022/botcommands/api/pagination/menu/buttonized/ChoiceCallback.kt
@@ -1,11 +1,11 @@
-package io.github.freya022.botcommands.api.pagination.menu
+package io.github.freya022.botcommands.api.pagination.menu.buttonized
 
 import io.github.freya022.botcommands.api.components.event.ButtonEvent
 
 typealias SuspendingChoiceCallback<E> = suspend (event: ButtonEvent, entry: E) -> Unit
 
 /**
- * Callback called when an item has been chosen in a [ChoiceMenu]
+ * Callback called when an item has been chosen in a [ButtonMenu]
  *
  * @param E Type of the entries
  */
@@ -14,7 +14,7 @@ fun interface BlockingChoiceCallback<E> {
      * Runs the callback
      *
      * @param event The [ButtonEvent] from the interacting user
-     * @param entry The selected entry from the [ChoiceMenu] elements
+     * @param entry The selected entry from the [ButtonMenu] elements
      */
     fun accept(event: ButtonEvent, entry: E)
 }
